[FEAT] 스웨거 어노테이션 추가 및 검색 기능 오류 해결

+++ src/main/java/com/sparta/quizdemo/backoffice/service/BackofficeService.java
@@ -5,7 +5,9 @@ import com.sparta.quizdemo.backoffice.repository.BackofficeRepository;
 import com.sparta.quizdemo.common.dto.ApiResponseDto;
 import com.sparta.quizdemo.common.entity.User;
 import com.sparta.quizdemo.user.UserRepository;
+import com.sparta.quizdemo.user.UserRequestDto;
 import com.sparta.quizdemo.user.UserResponseDto;
+import com.sparta.quizdemo.user.UserRoleEnum;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +16,7 @@ import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.HandlerInterceptor;
 
@@ -28,6 +31,7 @@ public class BackofficeService implements HandlerInterceptor {
 
     private final UserRepository userRepository;
     private final BackofficeRepository backofficeRepository;
+    private final PasswordEncoder passwordEncoder;
     private final RedisTemplate<String, String> redisTemplate;
 
     public ResponseEntity<List<Visitor>> getVisitors() {
@@ -88,10 +92,29 @@ public class BackofficeService implements HandlerInterceptor {
         return ResponseEntity.status(HttpStatus.OK).body(userResponseDtoList);
     }
 
-    public ResponseEntity<ApiResponseDto> deleteOneUser(Long userNo) {
-        User user = userRepository.findById(userNo).orElseThrow(() -> new NullPointerException("해당 번호의 유저가 존재하지 않습니다."));
-        userRepository.delete(user);
-        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("해당 번호의 유저를 탈퇴시켰습니다.", HttpStatus.OK.value()));
+    public ResponseEntity<UserResponseDto> updateOneUSer(String userName, UserRequestDto userRequestDto) {
+        User user = userRepository.findByUsername(userName).orElseThrow(() -> new NullPointerException("해당 ID의 유저가 존재하지 않습니다."));
+
+        if (user.getRole().equals(UserRoleEnum.ADMIN)) {
+            throw new IllegalArgumentException("관리자 권한을 가진 유저입니다.");
+        } else {
+            String newPassword = passwordEncoder.encode(userRequestDto.getNewPassword());
+            user.update(userRequestDto, newPassword);
+            userRepository.save(user);
+
+            return ResponseEntity.status(HttpStatus.OK).body(new UserResponseDto(user));
+        }
+    }
+
+    public ResponseEntity<ApiResponseDto> deleteOneUser(String userName) {
+        User user = userRepository.findByUsername(userName).orElseThrow(() -> new NullPointerException("해당 ID의 유저가 존재하지 않습니다."));
+
+        if (user.getRole().equals(UserRoleEnum.ADMIN)) {
+            throw new IllegalArgumentException("관리자 권한을 가진 유저입니다.");
+        } else {
+            userRepository.delete(user);
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("해당 ID의 유저를 탈퇴시켰습니다.", HttpStatus.OK.value()));
+        }
     }
 
     // 방문자 정보를 가져오기 위한 인터셉터 설정
